// src/lib/notion/api-types.ts
export type Block =
  | ParagraphBlock
  | HeadingOneBlock
  | HeadingTwoBlock
  | HeadingThreeBlock
  | BulletedListItemBlock
  | NumberedListItemBlock
  | ToDoBlock
  | ToggleBlock
  | ChildPageBlock
  | UnsupportedBlock

export interface BlockBase {
  object: 'block'
  id: string
  type: string
  created_time: string
  last_edited_time: string
  has_children: boolean
}

export interface ParagraphBlock extends BlockBase {
  type: 'paragraph'
  paragraph: {
    text: RichText[]
    children?: Block[]
  }
}

export interface HeadingOneBlock extends BlockBase {
  type: 'heading_1'
  heading_1: { text: RichText[] }
  has_children: false
}

export interface HeadingTwoBlock extends BlockBase {
  type: 'heading_2'
  heading_2: { text: RichText[] }
  has_children: false
}

export interface HeadingThreeBlock extends BlockBase {
  type: 'heading_3'
  heading_3: { text: RichText[] }
  has_children: false
}

export interface BulletedListItemBlock extends BlockBase {
  type: 'bulleted_list_item'
  bulleted_list_item: {
    text: RichText[]
    children?: Block[]
  }
}

export interface NumberedListItemBlock extends BlockBase {
  type: 'numbered_list_item'
  numbered_list_item: {
    text: RichText[]
    children?: Block[]
  }
}

export interface ToDoBlock extends BlockBase {
  type: 'to_do'
  to_do: {
    text: RichText[]
    checked: boolean
    children?: Block[]
  }
}

export interface ToggleBlock extends BlockBase {
  type: 'toggle'
  toggle: {
    text: RichText[]
    children?: Block[]
  }
}

export interface ChildPageBlock extends BlockBase {
  type: 'child_page'
  child_page: { title: string }
}

export interface UnsupportedBlock extends BlockBase {
  type: 'unsupported'
}
// ...
export type RichText = RichTextText | RichTextMention | RichTextEquation